requirements.txt
numpy
pandas
scikit-learn
xgboost
yfinance
matplotlib
seaborn
ta
torch
transformers
fastapi
uvicorn
python-multipart
streamlit
plotly
requests
python-dotenv



